tg-bot/requirements.txt
numpy
Pillow
matplotlib
ultralytics
asyncio
aiogram==3.7.0
python-dotenv